createa a shared serializer base for autofixTuning and stoppingPoint
