+++ src/sentry/seer/autofix/utils.py
@@ -8,6 +8,7 @@ import pydantic
 import requests
 from django.conf import settings
 from pydantic import BaseModel
+from rest_framework import serializers
 from urllib3 import Retry
 
 from sentry import features, options, ratelimits
@@ -140,6 +141,28 @@ class CodingAgentStateUpdateRequest(BaseModel):
 autofix_connection_pool = connection_from_url(
     settings.SEER_AUTOFIX_URL,
 )
+
+
+class SeerAutofixSettingsSerializer(serializers.Serializer):
+    """Base serializer for autofixAutomationTuning and automatedRunStoppingPoint"""
+
+    autofixAutomationTuning = serializers.ChoiceField(
+        choices=[opt.value for opt in AutofixAutomationTuningSettings],
+        required=False,
+        help_text="The tuning setting for the projects.",
+    )
+    automatedRunStoppingPoint = serializers.ChoiceField(
+        choices=[opt.value for opt in AutofixStoppingPoint],
+        required=False,
+        help_text="The stopping point for the projects.",
+    )
+
+    def validate(self, data):
+        if "autofixAutomationTuning" not in data and "automatedRunStoppingPoint" not in data:
+            raise serializers.ValidationError(
+                "At least one of 'autofixAutomationTuning' or 'automatedRunStoppingPoint' must be provided."
+            )
+        return data
 
 
 def get_project_seer_preferences(project_id: int) -> SeerRawPreferenceResponse:
